+++ packages/shared/forks/ReactFeatureFlags.native-fabric-fb.js
@@ -22,7 +22,6 @@ export const warnAboutLegacyContextAPI = __DEV__;
 export const replayFailedUnitOfWorkWithInvokeGuardedCallback = __DEV__;
 export const enableProfilerTimer = __PROFILE__;
 export const enableInteractionTracking = false;
-export const enableInteractionTrackingObserver = false;
 
 // Only used in www builds.
 export function addUserTimingListener() {
