+++ packages/shared/forks/ReactFeatureFlags.native-fb.js
@@ -27,7 +27,6 @@ export const enableUserTimingAPI = __DEV__;
 export const warnAboutLegacyContextAPI = __DEV__;
 export const enableProfilerTimer = __PROFILE__;
 export const enableInteractionTracking = false;
-export const enableInteractionTrackingObserver = false;
 
 // Only used in www builds.
 export function addUserTimingListener() {
