+++ packages/shared/forks/ReactFeatureFlags.test-renderer.js
@@ -22,7 +22,6 @@ export const warnAboutLegacyContextAPI = false;
 export const replayFailedUnitOfWorkWithInvokeGuardedCallback = false;
 export const enableProfilerTimer = false;
 export const enableInteractionTracking = false;
-export const enableInteractionTrackingObserver = false;
 
 // Only used in www builds.
 export function addUserTimingListener() {
